[propagating revision 4543 from development stream] Resolve gcc-4.4 and 64-bit warnings. (dm)
git-svn-id: svn://mielke.cc/branches/brltty-4.1@4823 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/ttb_louis.c
@@ -62,7 +62,7 @@ getUnicodeCharacters (DataFile *file, wchar_t *character, int num, const char *d
 	  default:
           invalid:
 	    reportDataError(file, "unknown escape sequence: %.*" PRIws,
-                            c-string.characters+1, string.characters);
+                            (int)(c-string.characters+1), string.characters);
 	}
       } else {
 	character[i] = *c++;
